Remove code that seems redundant.
When we start off with an aliasing pair returned by GetInPlaceInputOutputPairs,
we can be sure that the buffer will be shared. BufferAssignment does not use
the CanShareBuffer hook to disallow any in-place rule.
Besides, the call of the hook had another issue, it passed
operand_index.operand_index as value for user_index.

PiperOrigin-RevId: 773556820